fix: move logic from unused context factory to program
